requirements.txt
pandas>=1.3.0
requests>=2.26.0
Pillow>=8.3.1
pytesseract>=0.3.8
pdf2image>=1.16.0
PyPDF2>=2.0.0
python-dotenv>=0.19.0
openai>=1.0.0 